Use new API endpoint
This is the same API that is used by the Android application that is used by Huum. It supports the same authentication method as before as well.

Some of the response structure has changed with this update and the schemas are therefor changes. The main breaking change is the removal of `max_heating_time` in the reponse (a value that is now found in the `sauna_config`.

Note that this minor breaking change should not impact integrations such as Home Assistant, as that value is not used there.