Remove `status_from_status_or_stop`
This is a major breaking change, and will break the Home Assistant integration if code is not updated there.

The function is removed as the hack for getting the target temperature is no longer needed due to changes in the Huum API. The API now always returns the correct target temperature without needing to call the `off` endpoint.